+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.1-beta] - 2025-04-23
+
+### Added
+- CLI flags: `--qb_color`, `--qb_gray-color`, and `--qb_gray` for custom quantization bitness of color, grayscale+one, and grayscale images.
+- Conversion logic updated to use these flags if provided.
+- README updated to document new flags and usage.
+
 ## [v1.0-beta] - 2025-04-23
 
 ### Added